rename: check-duplicate 파일 shared로 이동 (#206)
